refactor(moot): only clean, no confusion
